[dev-client][iOS] autodetect launcher packager
Goal: make a small reduction in developer friction when generating test apps to work
on dev-client, dev-launcher, etc.

Currently, an Expo developer must make manual code changes in order for a test app
to load the dev-launcher packager (on port 8090 by default). This PR makes that unnecessary.
Additionally, the iOS launcher code will now gracefully detect whether the dev-launcher
packager is running, and fall back to using the prebuilt resource bundles if not.

After this change, if the developer sets the environment variable EX_DEV_LAUNCHER_URL
(normally to http://localhost:8090), `pod install` will set this as a define in the Xcode
workspace for the test app.  Code in  `EXDevLauncherController.m` checks the `status` endpoint
of the dev-launcher packager to see if it is running, and uses that to decide whether to load
from packager or bundles.

In addition to the iOS code changes, this includes a change in `autolinking.rb` used by `pod install`,
and a change in the `expo-template-bare-minimum` template Podfile.

Followed the steps above locally and verified that the generated test app works, and loads the packager
or the bundle appropriately.

- [ ] Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [ ] This diff will work correctly for `expo build` (eg: updated `@expo/xdl`).
- [ ] This diff will work correctly for `expo prebuild` & EAS Build (eg: updated a module plugin).